talk/pyproject.toml
[project]
name = "pretalx"
dynamic = ["version"]
description = "Conference organisation: CfPs, scheduling, much more"
readme = "README.rst"
requires-python = ">=3.11"
keywords = ["conference", "event", "schedule", "talk", "cfp"]
authors = [{ name = "Tobias Kunze", email = "r@rixx.de" }]
maintainers = [{ name = "Tobias Kunze", email = "r@rixx.de" }]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "Intended Audience :: Other Audience",
    "License :: OSI Approved :: Apache Software License",
    "License :: OSI Approved",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

dependencies = [
    "beautifulsoup4[lxml]~=4.13.0",
    "bleach~=6.2.0",
    "celery~=5.5.0",
    "css_inline~=0.14.0",
    "csscompressor~=0.9.0",
    "cssutils~=2.11.0",
    "defusedcsv~=2.0.0",
    "defusedxml~=0.7.0",
    "Django[argon2]~=5.2.2",
    "django-compressor~=4.5.0",
    "django-context-decorator",
    "django-countries~=7.0",
    "django-csp~=3.8.0",
    "django-filter==25.1",
    "django-formset-js-improved==0.5.0.3",
    "django-formtools~=2.5.1",
    "django-hierarkey~=1.2.0",
    "django-i18nfield>=1.9,<1.11",
    "django-libsass~=0.8",
    "django-scopes~=2.0.0",
    "django-pdb~=0.6.2",
    "djangorestframework~=3.16.0",
    "drf-flex-fields~=1.0.0",
    "drf-spectacular~=0.28.0",
    "libsass~=0.23.0",
    "Markdown~=3.8.0",
    "Pillow~=11.2.0",
    "publicsuffixlist~=1.0.0",
    "python-dateutil~=2.9.0",
    "qrcode~=8.0",
    "reportlab~=4.4.0",
    "requests~=2.32.0",
    "rules~=3.5.0",
    "setuptools",
    "urlman~=2.0.1",
    "vobject~=0.9.0",
    "whitenoise~=6.9.0",
    "zxcvbn~=4.5.0",
    "pretalx_pages @ git+https://github.com/hongquan/eventyay-talk-pages.git@2956f13",
    "pretalx_venueless @ git+https://github.com/fossasia/eventyay-talk-video.git@main",
    "pretalx-downstream @ git+https://github.com/fossasia/eventyay-talk-downstream.git@main",
    "django-allauth~=0.63.3",
    "PyJWT~=2.8.0",
    "django-oauth-toolkit==2.4.*",
    "requests-oauthlib==2.*",
    "psycopg2-binary==2.9.10",
    "oauthlib==3.2.*",
    "gunicorn>=23.0.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary~=2.9.0"]
redis = ["redis~=5.2.0"]
memcached = ["pylibmc"]

[project.urls]
Homepage = "https://eventyay.com"
Documentation = "https://docs.eventyay.org"
Repository = "https://github.com/fossasia/eventyay-talk"
Changelog = "https://docs.eventyay.org/changelog/"
Issues = "https://github.com/fossasia/eventyay-talk/issues"

[dependency-groups]
lint = ["djhtml>=3.0.7", "ruff>=0.9.10"]
test = [
    "coverage>=7.6.12",
    "freezegun>=1.5.1",
    "pytest>=8.3.5",
    "pytest-cov",
    "pytest-django",
    "pytest-mock",
    "pytest-sugar",
    "pytest-xdist",
    "responses>=0.25.7",
]
dev = [
    "build>=1.2.2.post1",
    "check-manifest>=0.50",
    "jsonschema>=4.23.0",
    "lxml>=5.3.1",
    "pywatchman>=2.0.0",
    "urllib3>=2.3.0",
]
doc = [
    "sphinx-autobuild>=2024.10.3",
    "sphinx>=8.2.3",
    "sphinx-autodoc-typehints>=3.1.0",
    "sphinx-copybutton>=0.5.2",
    "sphinxcontrib-django>=2.5",
    "sphinxcontrib-httpdomain>=1.8.1",
    "sphinxcontrib-spelling>=8.0.1",
]

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools", "wheel"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic.version]
attr = "pretalx.__version__"

[tool.setuptools.packages.find]
where = ["src"]
include = ["pretalx*"]
exclude = ["pretalx/frontend/node_modules"]
namespaces = false

[tool.check-manifest]
ignore = [
    ".*",
    "assets/*",
    "CODE_OF_CONDUCT.md",
    "CONTRIBUTORS",
    "docs/*",
    "src/.watchmanconfig",
    "src/local/*",
    "src/tests",
    "src/tests/*",
    "SECURITY.md",
    "justfile",
]

[tool.djlint]
blank_line_after_tag = "load,extends,include"
blank_line_before_tag = "load,extends,include"
close_void_tags = false
extend_exclude = "static,docs"
max_blank_lines = 1
preserve_blank_lines = 1
profile = "django"

[tool.uv]
default-groups = ["dev", "test"]
